requirements.txt
# DocuGen Requirements
# Core dependencies for Phase 1

# CLI framework
click>=8.0.0

# Data processing
pandas>=1.5.0
openpyxl>=3.0.0
python-dateutil>=2.8.0

# Template processing
python-docx>=0.8.11
Jinja2>=3.1.0

# PDF conversion
docx2pdf>=0.1.8

# Configuration
PyYAML>=6.0

# Progress tracking
tqdm>=4.64.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0